requirements.txt
fastapi==0.110.0
pillow==10.2.0
psycopg2==2.9.9
pydantic==2.6.4
python-dotenv==1.0.1
uvicorn==0.29.0